refactor(http-stress-test): cleanup
